+++ sc_headlines.php
@@ -28,7 +28,7 @@
 if(isset($rubricID) and $rubricID) $only = "AND rubric='".$rubricID."'";
 else $only='';
 
-$ergebnis=safe_query("SELECT * FROM ".PREFIX."news WHERE published='1' ".$only." AND intern<=".isclanmember($userID)." ORDER BY date DESC LIMIT 0,".$maxheadlines);
+$ergebnis=safe_query("SELECT * FROM ".PREFIX."news WHERE published='1' ".$only." AND intern<=".(int)isclanmember($userID)." ORDER BY date DESC LIMIT 0,".$maxheadlines);
 if(mysqli_num_rows($ergebnis)){
 	echo '<ul class="nav nav-pills">';
 	$n=1;
@@ -36,7 +36,7 @@ if(mysqli_num_rows($ergebnis)){
 		$date=getformatdate($ds['date']);
 		$time=getformattime($ds['date']);
 		$news_id=$ds['newsID'];
-		
+
 		if($n%2) {
 			$bg1=BG_1;
 			$bg2=BG_2;
@@ -45,35 +45,35 @@ if(mysqli_num_rows($ergebnis)){
 			$bg1=BG_3;
 			$bg2=BG_4;
 		}
-		
+
 		$message_array = array();
 		$query=safe_query("SELECT n.*, c.short AS `countryCode`, c.country FROM ".PREFIX."news_contents n LEFT JOIN ".PREFIX."countries c ON c.short = n.language WHERE n.newsID='".$ds['newsID']."'");
 		while($qs = mysqli_fetch_array($query)) {
 			$message_array[] = array('lang' => $qs['language'], 'headline' => $qs['headline'], 'message' => $qs['content'], 'country'=> $qs['country'], 'countryShort' => $qs['countryCode']);
 		}
 		$showlang = select_language($message_array);
-	  
+
 		$languages='';
 		$i=0;
 		foreach($message_array as $val) {
 			if($showlang!=$i)	$languages.='<span style="padding-left:2px"><a href="index.php?site=news_comments&amp;newsID='.$ds['newsID'].'&amp;lang='.$val['lang'].'"><img src="images/flags/'.$val['countryShort'].'.gif" width="18" height="12" alt="'.$val['country'].'"></a></span>';
 			$i++;
 		}
-	  
+
 		$lang=$message_array[$showlang]['lang'];
-	
+
 		$headlines=$message_array[$showlang]['headline'];
-	
+
 		if(mb_strlen($headlines)>$maxheadlinechars) {
 			$headlines=mb_substr($headlines, 0, $maxheadlinechars);
 			$headlines.='...';
 		}
-	
+
 		$headlines=clearfromtags($headlines);
-	
+
 		eval ("\$sc_headlines = \"".gettemplate("sc_headlines")."\";");
 		echo $sc_headlines;
-		
+
 		$n++;
 	}
 	echo '</ul>';
